Refactor: BallNumber, BallPosition
BallNumber: 함수 명 isSame -> isSameNumber 변경
BallNumber, BallPosition: 패키지 이동
