+++ src/main/java/baseball/domain/ball/ballcomponent/BallNumber.java
@@ -1,4 +1,4 @@
-package baseball.domain.ball;
+package baseball.domain.ball.ballcomponent;
 
 import java.util.Objects;
 
@@ -9,7 +9,7 @@ public class BallNumber {
         this.ballNumber = ballNumber;
     }
 
-    public boolean isSame(BallNumber ballNumber2) {
+    public boolean isSameNumber(BallNumber ballNumber2) {
         return Objects.equals(this.ballNumber, ballNumber2.ballNumber);
     }
 }
